pyproject.toml
[project]
name = "gnn-from-scratch"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "datasets (>=3.3.2,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "torchinfo (>=1.8.0,<2.0.0)",
    "wandb (>=0.19.7,<0.20.0)",
    "torch-geometric (>=2.6.1,<3.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "torchmetrics (>=1.7.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

